+++ ijs/IElement.cs
@@ -4,5 +4,5 @@ public interface IElement : IEquatable<IElement> {
     string Name { get; }
     string FullCode { get; }
     ElementType ElementType { get; }
-    float BaseValue { get; }
+    double BaseValue { get; }
 }
