+++ ijs/SecondHalfElement.cs
@@ -1,7 +1,7 @@
 ﻿namespace ijs;
 
 public class SecondHalfElement : IElement {
-    public const float Magnification = 1.1F;
+    public const double Magnification = 1.1F;
     readonly IElement element;
 
     public SecondHalfElement(IElement element) {
@@ -15,5 +15,5 @@ public class SecondHalfElement : IElement {
     public string Name => element.Name;
     public string FullCode => element.FullCode;
     public ElementType ElementType => element.ElementType;
-    public float BaseValue => element.BaseValue * Magnification;
+    public double BaseValue => element.BaseValue * Magnification;
 }
